agrovision-azure-func/iot/requirements.txt
azure-functions
azure-iot-hub
azure-cosmos
requests